+++ script.audio.motherearth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1
+- updated to https 
+- added artist fanart background when available
+
 ## v2.0
 
 - metadata for the streams